requirements.txt
Django>=5.2,<6.0

gunicorn>=22
whitenoise>=6.7
dj-database-url>=2.2
python-dotenv>=1.0

requests>=2.32

django-allauth>=0.63

celery>=5.4
redis>=5.0
django-celery-beat>=2.6

psycopg2-binary>=2.9.10

# 你目前有 channels / uvicorn 就先留著（要用 websocket 再說）
channels>=4.2
channels-redis>=4.2
uvicorn>=0.30

PyJWT>=2.8.0
cryptography>=43
